requirements.txt
pyyaml>=6.0
jsonschema>=4.19.0
pytest>=7.4.0